Move venv to workspace root
This venv was initially used only for ansible so was stored in playbooks
Now its used more widely, so its moved out of playbooks into the
workspace root.

This commit also removes venv as a variable to some functions to
show that there is only one venv that is always used.